tests/wpt/web-platform-tests/event-timing/programmatic-click-not-observed.html

<!DOCTYPE html>
<html>
<meta charset=utf-8 />
<div id='div' onclick='delay()'>Click me</div>
<div id='div2'>No, click me!</div>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/resources/testdriver.js></script>
<script src=/resources/testdriver-vendor.js></script>

<script src=resources/event-timing-test-utils.js></script>
<script>
  let delayCalled = false;
  let beforeClick;
  function delay() {
    const end = performance.now() + 150;
    while(performance.now() < end) {}
    delayCalled = true;
  }
  async_test(function(t) {
    if (!window.PerformanceEventTiming)
      assert_unreached("PerformanceEventTiming is not supported");

    const observer = new PerformanceObserver(t.step_func_done((entryList) => {
      const entries = entryList.getEntries().filter(e => e.name === 'mousedown');
      // There must only be one click entry: from the clickAndBlockMain() call.
      assert_equals(entries.length, 1);
      const entry = entries[0];
      // This ensures that the entry is exposing timing from the second click, i.e.
      // the one from the clickAndBlockMain() call.
      assert_greater_than_equal(entry.processingStart, beforeClick);
      // Check that the first input entry was also from the second click.
      const firstInput = performance.getEntriesByType('first-input');
      assert_equals(firstInput.length, 1);
      assert_greater_than_equal(firstInput[0].processingStart, beforeClick);
    }));
    observer.observe({entryTypes: ['event']});
    document.getElementById('div').click();
    // Take the timestamp after the programmatic click but before the next click.
    beforeClick = performance.now();
    // After the programmatic click, use another input to know when entries have been
    // dispatched to the PerformanceObserver callback.
    clickAndBlockMain('div2');
  }, "Event Timing: events from programmatic click are not observed");
</script>
</html>
