LayoutTests/imported/w3c/web-platform-tests/event-timing/programmatic-click-not-observed-expected.txt

Click me
No, click me!

FAIL Event Timing: events from programmatic click are not observed assert_equals: There should be a single first-input entry expected 1 but got 0
